+++ Knjigovodstvo/JoppdDocument/JoppdEntitet.cs
@@ -22,7 +22,7 @@ namespace Knjigovodstvo.JoppdDocument
                                   Id = int.Parse(dr["Id"].ToString()),
                                   Oib = dr["Oib"].ToString(),
                                   Sifra = dr["Sifra"].ToString(),
-                                  Iznos = float.Parse(dr["Iznos"].ToString())
+                                  Iznos = decimal.Parse(dr["Iznos"].ToString())
                               }).ToList();
 
             if (_dodaci.Count > 0)
@@ -69,51 +69,51 @@ namespace Knjigovodstvo.JoppdDocument
         [XmlElement("P102")]
         public string Datum_Do { get; set; } = "0";
         [XmlElement("P11")]
-        public float Bruto { get; set; } = 0;
+        public decimal Bruto { get; set; } = 0;
         [XmlElement("P12")]
-        public float Osnovica { get; set; } = 0;
+        public decimal Osnovica { get; set; } = 0;
         [XmlElement("P121")]
-        public float Mio_1 { get; set; } = 0;
+        public decimal Mio_1 { get; set; } = 0;
         [XmlElement("P122")]
-        public float Mio_2 { get; set; } = 0;
+        public decimal Mio_2 { get; set; } = 0;
         [XmlElement("P123")]
-        public float Zdravstvo { get; set; } = 0;
+        public decimal Zdravstvo { get; set; } = 0;
         [XmlElement("P124")]
-        public float Zaštita { get; set; } = 0;
+        public decimal Zaštita { get; set; } = 0;
         [XmlElement("P125")]
-        public float Zaposljavanje { get; set; } = 0;
+        public decimal Zaposljavanje { get; set; } = 0;
         [XmlElement("P126")]
-        public float Povecani_Mio { get; set; } = 0;
+        public decimal Povecani_Mio { get; set; } = 0;
         [XmlElement("P127")]
-        public float Povecani_Mio2 { get; set; } = 0;
+        public decimal Povecani_Mio2 { get; set; } = 0;
         [XmlElement("P128")]
-        public float Zdravstvo_Inozemstvo { get; set; } = 0;
+        public decimal Zdravstvo_Inozemstvo { get; set; } = 0;
         [XmlElement("P129")]
-        public float Invaliditet_Doprinos { get; set; } = 0;
+        public decimal Invaliditet_Doprinos { get; set; } = 0;
         [XmlElement("P131")]
-        public float Izdatak { get; set; } = 0;
+        public decimal Izdatak { get; set; } = 0;
         [XmlElement("P132")]
-        public float IzdatakUplaceni_Mio { get; set; } = 0;
+        public decimal IzdatakUplaceni_Mio { get; set; } = 0;
         [XmlElement("P133")]
-        public float Dohodak { get; set; } = 0;
+        public decimal Dohodak { get; set; } = 0;
         [XmlElement("P134")]
-        public float Osobni_Odbitak { get; set; } = 0;
+        public decimal Osobni_Odbitak { get; set; } = 0;
         [XmlElement("P135")]
-        public float Porezna_Osnovica { get; set; } = 0;
+        public decimal Porezna_Osnovica { get; set; } = 0;
         [XmlElement("P141")]
-        public float Porez_Ukupno { get; set; } = 0;
+        public decimal Porez_Ukupno { get; set; } = 0;
         [XmlElement("P142")]
-        public float Prirez { get; set; } = 0;
+        public decimal Prirez { get; set; } = 0;
         [XmlElement("P151")]
         public string Oznaka_Neoporezivog { get; set; } = "0";
         [XmlElement("P152")]
-        public float Iznos_Neoporezivog { get; set; } = 0;
+        public decimal Iznos_Neoporezivog { get; set; } = 0;
         [XmlElement("P161")]
         public string Nacin_Isplate { get; set; } = "0";
         [XmlElement("P162")]
-        public float Iznos_Isplate { get; set; } = 0;
+        public decimal Iznos_Isplate { get; set; } = 0;
         [XmlElement("P17")]
-        public float Primitak_Nesamostalni { get; set; } = 0;        
+        public decimal Primitak_Nesamostalni { get; set; } = 0;        
     }
 
     [Serializable()]
