+++ Knjigovodstvo/Payroll/Placa.cs
@@ -13,10 +13,10 @@ namespace Knjigovodstvo.Payroll
             return FormError.None;
         }
 
-        public Placa Izracun(float bruto, float prirez, float dodaci, float odbitak, bool drugi = false)
+        public Placa Izracun(decimal bruto, decimal prirez, decimal dodaci, decimal odbitak, bool drugi = false)
         {
             Postavke p = new Postavke();
-            float iznos = bruto;
+            decimal iznos = bruto;
             Bruto = iznos;
             if (drugi)
             {
@@ -30,12 +30,12 @@ namespace Knjigovodstvo.Payroll
                 iznos -= Mio_1 + Mio_2;
             }
             Dohodak = iznos;
-            iznos -= Osobni_Odbitak = 2500.0f * (1.6f + odbitak);
+            iznos -= Osobni_Odbitak = 2500.0m * (1.6m + odbitak);
             Porezna_Osnovica = iznos;
 
             if (Porezna_Osnovica > 30000)
             {
-                iznos -= Porez_24_per = 30000.0f * p.GetStopaByName(PlacaStope.Porez_Dohodak_24);
+                iznos -= Porez_24_per = 30000.0m * p.GetStopaByName(PlacaStope.Porez_Dohodak_24);
                 iznos -= Porez_36_per = (Porezna_Osnovica - 30000) * p.GetStopaByName(PlacaStope.Porez_Dohodak_36);
             }
             else
@@ -61,20 +61,20 @@ namespace Knjigovodstvo.Payroll
             {
                 Id = int.Parse(data.Rows[0]["Id"].ToString());
                 Oib = data.Rows[0]["Oib"].ToString();
-                Bruto = float.Parse(data.Rows[0]["Bruto"].ToString());
-                Mio_1 = float.Parse(data.Rows[0]["Mio_1"].ToString());
-                Mio_2 = float.Parse(data.Rows[0]["Mio_2"].ToString());
-                Dohodak = float.Parse(data.Rows[0]["Dohodak"].ToString());
-                Osobni_Odbitak = float.Parse(data.Rows[0]["Osobni_Odbitak"].ToString());
-                Porezna_Osnovica = float.Parse(data.Rows[0]["Porezna_Osnovica"].ToString());
-                Porez_24_per = float.Parse(data.Rows[0]["Porez_24_per"].ToString());
-                Porez_36_per = float.Parse(data.Rows[0]["Porez_36_per"].ToString());
-                Porez_Ukupno = float.Parse(data.Rows[0]["Porez_Ukupno"].ToString());
-                Prirez = float.Parse(data.Rows[0]["Prirez"].ToString());
-                Ukupno_Porez_i_Prirez = float.Parse(data.Rows[0]["Ukupno_Porez_i_Prirez"].ToString());
-                Neto = float.Parse(data.Rows[0]["Neto"].ToString());
-                Doprinos_Zdravstvo = float.Parse(data.Rows[0]["Doprinos_Zdravstvo"].ToString());
-                Dodaci_Ukupno = float.Parse(data.Rows[0]["Dodaci_Ukupno"].ToString());
+                Bruto = decimal.Parse(data.Rows[0]["Bruto"].ToString());
+                Mio_1 = decimal.Parse(data.Rows[0]["Mio_1"].ToString());
+                Mio_2 = decimal.Parse(data.Rows[0]["Mio_2"].ToString());
+                Dohodak = decimal.Parse(data.Rows[0]["Dohodak"].ToString());
+                Osobni_Odbitak = decimal.Parse(data.Rows[0]["Osobni_Odbitak"].ToString());
+                Porezna_Osnovica = decimal.Parse(data.Rows[0]["Porezna_Osnovica"].ToString());
+                Porez_24_per = decimal.Parse(data.Rows[0]["Porez_24_per"].ToString());
+                Porez_36_per = decimal.Parse(data.Rows[0]["Porez_36_per"].ToString());
+                Porez_Ukupno = decimal.Parse(data.Rows[0]["Porez_Ukupno"].ToString());
+                Prirez = decimal.Parse(data.Rows[0]["Prirez"].ToString());
+                Ukupno_Porez_i_Prirez = decimal.Parse(data.Rows[0]["Ukupno_Porez_i_Prirez"].ToString());
+                Neto = decimal.Parse(data.Rows[0]["Neto"].ToString());
+                Doprinos_Zdravstvo = decimal.Parse(data.Rows[0]["Doprinos_Zdravstvo"].ToString());
+                Dodaci_Ukupno = decimal.Parse(data.Rows[0]["Dodaci_Ukupno"].ToString());
             }
             catch
             {
@@ -86,19 +86,19 @@ namespace Knjigovodstvo.Payroll
 
         public int Id { get; set; } = 0;
         public string Oib { get; set; } = "";
-        public float Bruto { get; set; } = 0;
-        public float Mio_1 { get; set; } = 0;
-        public float Mio_2 { get; set; } = 0;
-        public float Dohodak { get; set; } = 0;
-        public float Osobni_Odbitak { get; set; } = 0;
-        public float Porezna_Osnovica { get; set; } = 0;
-        public float Porez_24_per { get; set; } = 0;
-        public float Porez_36_per { get; set; } = 0;
-        public float Porez_Ukupno { get; set; } = 0;
-        public float Prirez { get; set; } = 0;
-        public float Ukupno_Porez_i_Prirez { get; set; } = 0;
-        public float Neto { get; set; } = 0;
-        public float Doprinos_Zdravstvo { get; set; } = 0;
-        public float Dodaci_Ukupno { get; set; } = 0;
+        public decimal Bruto { get; set; } = 0;
+        public decimal Mio_1 { get; set; } = 0;
+        public decimal Mio_2 { get; set; } = 0;
+        public decimal Dohodak { get; set; } = 0;
+        public decimal Osobni_Odbitak { get; set; } = 0;
+        public decimal Porezna_Osnovica { get; set; } = 0;
+        public decimal Porez_24_per { get; set; } = 0;
+        public decimal Porez_36_per { get; set; } = 0;
+        public decimal Porez_Ukupno { get; set; } = 0;
+        public decimal Prirez { get; set; } = 0;
+        public decimal Ukupno_Porez_i_Prirez { get; set; } = 0;
+        public decimal Neto { get; set; } = 0;
+        public decimal Doprinos_Zdravstvo { get; set; } = 0;
+        public decimal Dodaci_Ukupno { get; set; } = 0;
     }
 }
