+++ Knjigovodstvo/Payroll/PlacaObracun.cs
@@ -26,20 +26,20 @@ namespace Knjigovodstvo.Payroll
 
         public int Id { get; set; } = 0;
         public string Oib { get; set; } = "";
-        public float Bruto { get; set; } = 0;
-        public float Mio_1 { get; set; } = 0;
-        public float Mio_2 { get; set; } = 0;
-        public float Dohodak { get; set; } = 0;
-        public float Osobni_Odbitak { get; set; } = 0;
-        public float Porezna_Osnovica { get; set; } = 0;
-        public float Porez_24_per { get; set; } = 0;
-        public float Porez_36_per { get; set; } = 0;
-        public float Porez_Ukupno { get; set; } = 0;
-        public float Prirez { get; set; } = 0;
-        public float Ukupno_Porez_i_Prirez { get; set; } = 0;
-        public float Neto { get; set; } = 0;
-        public float Doprinos_Zdravstvo { get; set; } = 0;
-        public float Dodaci_Ukupno { get; set; } = 0;
+        public decimal Bruto { get; set; } = 0;
+        public decimal Mio_1 { get; set; } = 0;
+        public decimal Mio_2 { get; set; } = 0;
+        public decimal Dohodak { get; set; } = 0;
+        public decimal Osobni_Odbitak { get; set; } = 0;
+        public decimal Porezna_Osnovica { get; set; } = 0;
+        public decimal Porez_24_per { get; set; } = 0;
+        public decimal Porez_36_per { get; set; } = 0;
+        public decimal Porez_Ukupno { get; set; } = 0;
+        public decimal Prirez { get; set; } = 0;
+        public decimal Ukupno_Porez_i_Prirez { get; set; } = 0;
+        public decimal Neto { get; set; } = 0;
+        public decimal Doprinos_Zdravstvo { get; set; } = 0;
+        public decimal Dodaci_Ukupno { get; set; } = 0;
         public string Datum_Od { get; set; } = "";
         public string Datum_Do { get; set; } = "";
